requirements.txt
mock==4.0.2
msgpack==0.6.2
pymetasploit3==1.0.1
amqpstorm==2.8.1
pytest==6.1.0
pytest-cov==2.10.1
structlog==20.1.0
PyYAML==5.3.1
schema==0.7.2
click==7.1.2
